# Blur performance optimization