Hey just digging around for a faster blur and managed to trade a bit of memory for some speed.  Offering upstream in case you want the changes.

```
benchmark           old ns/op     new ns/op     delta
BenchmarkBlur-4     23625266      20647452      -12.60%

benchmark           old allocs     new allocs     delta
BenchmarkBlur-4     22             30             +36.36%

benchmark           old bytes     new bytes     delta
BenchmarkBlur-4     1961043       2097189       +6.94%
```