@disintegration 2018-05-18
Hello, 

I can confirm your benchmark results and I think it's a nice improvement. At the same time I'm not sure I understand the logic behind some of the changes. 

For example, you use subslice `v` of the scanLine in `blurVertical` but not in `blurHorlizontal`:

```
	v := scanLineF[i : i+4]
	wa := v[3] * weight
	r += v[0] * wa
	g += v[1] * wa
	b += v[2] * wa
```

 And another example - setting the calculated values to the `scanLine` slice first and copying it to the destination image after it (in `blurHorizontal`):

```
	scanLine[idx] = clamp(r)
	scanLine[idx+1] = clamp(g)
	scanLine[idx+2] = clamp(b)
	scanLine[idx+3] = clamp(a / wsum)
}
copy(dst.Pix[y*dst.Stride:], scanLine)
```

I have tested your changes one by one and it seems to me that most of the speedup is caused by creating an additional float64 slice `scanLineF` instead of converting ints to floats inside the innermost loop. Other changes don’t seem to add any noticeable improvement. 

Could you please try to revert all but the `scanlineF`-related changes and see if it’s the case for you too? I would like to merge the minimum set of changes that leads to the speedup.