@disintegration 2018-06-04
Thank you for the contribution! I'll try to further improve the blur performance based on your work.