@jasonmoo 2018-06-04
Incidentally the other changes you mentioned that seemed unrelated came up
as hotspots when I profiled the code but feel free to discard them if they
appear to be ineffective in the benchmarks.

On Mon, Jun 4, 2018 at 2:01 PM Jason Mooberry <jason.mooberry@gmail.com>
wrote:

> Ah yes!  Sorry I planned on rolling in your feedback and lost track of how
> long it had been. Glad you found this useful. Thanks for a great lib.
>
> On Mon, Jun 4, 2018 at 1:59 PM Grigory Dryapak <notifications@github.com>
> wrote:
>
>> Thank you for the contribution! I'll try to further improve the blur
>> performance based on your work.
>>
>> —
>> You are receiving this because you authored the thread.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/disintegration/imaging/pull/73#issuecomment-394444062>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AAO0eLu4W_N5kZi2SWnMHnm2EMDpjrlUks5t5XVpgaJpZM4T-g9b>
>> .
>>
>