@disintegration 2018-06-04
I've tested it again and made a minor change (dropped usage of subslice `v` and removed unused idx variable in `blurVertical`) and left other of your changes. Thanks again.